Split up an int test case.
